# backend/docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrueAdmin API",
    "version": "0.1.0",
    "description": "Generated from TrueAdmin controller attributes. Code annotations provide default interface metadata."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/admin/message-management/announcements": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminAnnouncementController_create",
        "summary": "创建公告",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminAnnouncementController@create",
          "permission": "system:announcement:create",
          "permissionMode": "single",
          "permissions": [
            "system:announcement:create"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminAnnouncementController_list",
        "summary": "公告列表",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminAnnouncementController@list",
          "permission": "system:announcement:list",
          "permissionMode": "single",
          "permissions": [
            "system:announcement:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/message-management/announcements/{id}": {
      "delete": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminAnnouncementController_delete",
        "summary": "删除公告草稿",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminAnnouncementController@delete",
          "permission": "system:announcement:delete",
          "permissionMode": "single",
          "permissions": [
            "system:announcement:delete"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminAnnouncementController_detail",
        "summary": "公告详情",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminAnnouncementController@detail",
          "permission": "system:announcement:detail",
          "permissionMode": "single",
          "permissions": [
            "system:announcement:detail"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminAnnouncementController_update",
        "summary": "编辑公告",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminAnnouncementController@update",
          "permission": "system:announcement:update",
          "permissionMode": "single",
          "permissions": [
            "system:announcement:update"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/message-management/announcements/{id}/cancel-scheduled": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminAnnouncementController_cancelScheduled",
        "summary": "取消定时公告",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminAnnouncementController@cancelScheduled",
          "permission": "system:announcement:update",
          "permissionMode": "single",
          "permissions": [
            "system:announcement:update"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/message-management/announcements/{id}/offline": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminAnnouncementController_offline",
        "summary": "下线公告",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminAnnouncementController@offline",
          "permission": "system:announcement:offline",
          "permissionMode": "single",
          "permissions": [
            "system:announcement:offline"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/message-management/announcements/{id}/publish": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminAnnouncementController_publish",
        "summary": "发布公告",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminAnnouncementController@publish",
          "permission": "system:announcement:publish",
          "permissionMode": "single",
          "permissions": [
            "system:announcement:publish"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/message-management/announcements/{id}/restore": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminAnnouncementController_restore",
        "summary": "恢复公告",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminAnnouncementController@restore",
          "permission": "system:announcement:update",
          "permissionMode": "single",
          "permissions": [
            "system:announcement:update"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/organization/departments": {
      "post": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminDepartmentController_create",
        "summary": "新增部门",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminDepartmentController@create",
          "permission": "system:department:create",
          "permissionMode": "single",
          "permissions": [
            "system:department:create"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminDepartmentController_list",
        "summary": "部门列表",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminDepartmentController@list",
          "permission": "system:department:list",
          "permissionMode": "single",
          "permissions": [
            "system:department:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/organization/departments/{id}": {
      "delete": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminDepartmentController_delete",
        "summary": "删除部门",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminDepartmentController@delete",
          "permission": "system:department:delete",
          "permissionMode": "single",
          "permissions": [
            "system:department:delete"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminDepartmentController_detail",
        "summary": "部门详情",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminDepartmentController@detail",
          "permission": "system:department:detail",
          "permissionMode": "single",
          "permissions": [
            "system:department:detail"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminDepartmentController_update",
        "summary": "编辑部门",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminDepartmentController@update",
          "permission": "system:department:update",
          "permissionMode": "single",
          "permissions": [
            "system:department:update"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/organization/departments/tree": {
      "get": {
        "operationId": "App_Module_System_Http_Admin_Controller_AdminDepartmentController_tree",
        "summary": "部门树",
        "description": "",
        "tags": [
          "系统管理"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-trueadmin": {
          "action": "App\\Module\\System\\Http\\Admin\\Controller\\AdminDepartmentController@tree",
          "permission": "system:department:list",
          "permissionMode": "single",
          "permissions": [
            "system:department:list"
          ],
          "public": false
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}